pyproject.toml
[project]
name = "hcmai2025-baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "beanie>=2.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=6.30.0",
    "jinja2>=3.1.6",
    "langdetect==1.0.9",
    "llama-index>=0.13.1",
    "llama-index-llms-google-genai>=0.3.0",
    "motor>=3.7.1",
    "nicegui>=2.22.1",
    "numpy>=2.2.6",
    "open-clip-torch>=3.0.0",
    "pydantic-settings>=2.10.1",
    "pymilvus>=2.5.14",
    "streamlit>=1.47.1",
    "torch>=2.7.1",
    "typing-extensions>=4.14.1",
    "usearch>=2.19.1",
    "uvicorn>=0.35.0",
    "deep-translator",
]
